requirements.txt
langchain
openai
requests
beautifulsoup4
streamlit
python-dotenv